# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tinder App API",
        "description": "API Documentation for Tinder-like Application",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/v1/people/{personId}/recommended": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get recommended people",
                "description": "Get list of recommended people with pagination. Excludes people that the current person has already interacted with.",
                "operationId": "554b9e899685078b30f0c47536a23639",
                "parameters": [
                    {
                        "name": "personId",
                        "in": "path",
                        "description": "ID of the current person",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Person not found"
                    }
                }
            }
        },
        "/api/v1/interactions/like": {
            "post": {
                "tags": [
                    "Interactions"
                ],
                "summary": "Like a person",
                "description": "Record that one person likes another person",
                "operationId": "79f12b9ac0c87f6f92a4148af66d6884",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "from_person_id",
                                    "to_person_id"
                                ],
                                "properties": {
                                    "from_person_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "to_person_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Person liked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Person liked successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/interactions/dislike": {
            "post": {
                "tags": [
                    "Interactions"
                ],
                "summary": "Dislike a person",
                "description": "Record that one person dislikes another person",
                "operationId": "65ca816273fc5f44a7ce016115ebb155",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "from_person_id",
                                    "to_person_id"
                                ],
                                "properties": {
                                    "from_person_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "to_person_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Person disliked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Person disliked successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/people/{personId}/liked-by": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get liked by list",
                "description": "Get list of people who liked the current person",
                "operationId": "cb4bcd260f605ac5fb2c4ba3f9dc1d72",
                "parameters": [
                    {
                        "name": "personId",
                        "in": "path",
                        "description": "ID of the person",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Person not found"
                    }
                }
            }
        },
        "/api/v1/people/{personId}/disliked-by": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get disliked by list",
                "description": "Get list of people who disliked the current person",
                "operationId": "ead988db8288690f88b368f3b976c225",
                "parameters": [
                    {
                        "name": "personId",
                        "in": "path",
                        "description": "ID of the person",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Person not found"
                    }
                }
            }
        },
...
    },
...
}